refactor(pipeline-odpt): split heuristic core out of build-timetable
Move Yurikamome-tuned trip-identity inference (preprocessTimetables,
estimateTravelTimes, inferTripsForDestination, buildCanonicalTrips,
countLayer1Origins) into infer-odpt-trips-heuristic.ts so the
heuristic's scope and assumptions are isolated from the generic
facade. build-timetable.ts becomes thin orchestration: dispatch to
canonical fast path / heuristic / legacy fallback, pattern
aggregation, and stopTimetable enrollment.

Test files mirror the split:
- build-odpt-timetable.test.ts covers the facade.
- infer-odpt-trips-heuristic.test.ts covers Issue #153 fixtures and
  heuristic helper unit tests.
- Shared scaffolding (makeOrder / makeRailway / makeTimetable) lives
  in __tests__/__fixtures__/odpt-timetable-fixtures.ts.

Pure structural refactor: yurimo data.json byte-identical pre/post
(md5 2b91a7e5566446e8846fc66a967618f7).

Refs: #153
